Handle review comments.
1. Remove lib/Semantics/check-omp-declare-variant.cpp and move the checks in check-omp-metadirective.cpp.

2. Move some functions into private from public  in OmpStructureChecker.

3. Update error for dynamic condition.
