+++ flang/test/Semantics/OpenMP/declare-variant-match.f90
@@ -76,7 +76,7 @@ end subroutine
 
 subroutine f06(x)
   integer :: x
-!ERROR: USER condition in the MATCH clause must be a constant expression
+!ERROR: Run-time USER condition in the MATCH clause is not yet implemented
   !$omp declare variant (sub:vsub) match (user={condition(x > 0)})
 contains
   subroutine vsub
